refactor(scripts): reorganize scripts/ into build/check/smoke subdirectories
Split the flat scripts/ root into role buckets:
- build/   release artifact producers (build-exe, build-assets, build-installer,
           build-npm-package, stage-windows-installer)
- check/   gates (check-npm-pack, check-staged-markdown-wrap, sync-bundled-skill-references)
- smoke/   end-to-end regression (CI-wired + manual PTY), with support/pty-smoke.ts
           moved into smoke/support/ beside its consumers
palette.ts stays at the root as the lone dev swatch utility.

Path bookkeeping:
- import.meta.dir root offset deepened by one '..' in 11 scripts that compute REPO_ROOT
- ../package.json -> ../../package.json in 3 scripts that import the manifest
- build-installer's internal call to build-exe updated to the new path
- package.json alias paths, release-build.yml (.ps1), git hook (via alias from
  the prior commit), contract test imports, and script usage docstrings updated
- docs/dev/AUDIT_DRIFT.md and STYLE.md path references + mirror synced

Verified: typecheck, lint (688 files), bun test (1999 pass / 0 fail).
