+++ .github/workflows/release-build.yml
@@ -164,7 +164,7 @@ jobs:
         shell: pwsh
         run: |
           bun run build:installer
-          ./scripts/smoke-windows-installer.ps1 -InstallerPath "dist/PrismVesicleSetup-*-windows-x64.exe"
+          ./scripts/smoke/smoke-windows-installer.ps1 -InstallerPath "dist/PrismVesicleSetup-*-windows-x64.exe"
       - name: Stage versioned Windows artifacts
         shell: pwsh
         env:
