+++ docs/dev/AUDIT_DRIFT.md
@@ -9,7 +9,7 @@ Three audit gates can go red from drift:
 | Gate | Location | Scope |
 |------|----------|-------|
 | `bun audit` | `.github/workflows/release-build.yml` (`checks` job, shared by PR/develop CI and the release pipeline) | Full dependency tree of the checkout |
-| Consumer `npm audit --omit=dev --audit-level=low` | `scripts/smoke-npm-package.ts` (`bun run pack:smoke`) | Production dependencies as installed by an npm consumer |
+| Consumer `npm audit --omit=dev --audit-level=low` | `scripts/smoke/smoke-npm-package.ts` (`bun run pack:smoke`) | Production dependencies as installed by an npm consumer |
 | Manual release verification | [`WORKFLOW.md`](./WORKFLOW.md) verification table | `bun audit` before a release tag |
 
 The vulnerability database is monotonic: advisories are added, essentially never removed. Rerunning a failed job (`gh run rerun`) cannot recover a drift red — the same database answer comes back. The retry rules in [`WORKFLOW.md`](./WORKFLOW.md) do not apply to audit failures.
@@ -45,7 +45,7 @@ Use this only when no patched release exists, or the patched release is breaking
 2. Add the finding to the `bun audit` invocation as `--ignore=<CVE>` (Bun ignores by CVE id; repeat the flag for multiple ids).
 3. Record the waiver in the commit message and in `CHANGELOG.md`: CVE id, justification from the triage, owner, and an explicit expiry/revisit date.
 4. Before the revisit date, re-triage: adopt a patch if one has appeared, renew with fresh justification, or drop the waiver. Expired waivers are never silently carried forward.
-5. The consumer `npm audit` gate has no ignore mechanism. Waiving a finding there requires changing `scripts/smoke-npm-package.ts` to parse `npm audit --json` and filter waived ids; that code change must be proposed and justified in the same PR as the waiver, and it stays limited to ids with a recorded, unexpired waiver.
+5. The consumer `npm audit` gate has no ignore mechanism. Waiving a finding there requires changing `scripts/smoke/smoke-npm-package.ts` to parse `npm audit --json` and filter waived ids; that code change must be proposed and justified in the same PR as the waiver, and it stays limited to ids with a recorded, unexpired waiver.
 
 ## Release Rule
