+++ host-assets/skills/vesicle-docs/references/dev-style.md
@@ -63,7 +63,7 @@ Flat directory layouts are discouraged once an area contains distinct domains, a
 - Keep directory roots for deliberate entry points, public facades, registries, and a small set of genuinely peer modules. Do not use a root as a catch-all because choosing a domain is inconvenient.
 - Avoid taxonomy theatre: a directory containing one trivial file or layers of pass-through re-exports is not useful modularization.
 - Shared directories such as `tests/support/` contain cross-domain infrastructure with a clear reusable contract, not miscellaneous helpers or fixtures that merely lack an owner.
-- Tests use a level-first, domain-second structure: `unit/`, `component/`, `integration/`, `contract/`, and `acceptance/`, followed by the subsystem under test. System and release smoke may live in `scripts/` and CI when that is the real execution boundary.
+- Tests use a level-first, domain-second structure: `unit/`, `component/`, `integration/`, `contract/`, and `acceptance/`, followed by the subsystem under test. System and release smoke may live in `scripts/smoke/` and CI when that is the real execution boundary; the `scripts/` tree splits into `build/`, `check/`, and `smoke/` roles.
 - Test fixtures belong under the nearest suite or domain that owns their meaning. Promote a fixture or helper to shared support only after independent domains genuinely reuse the same contract.
 
 A small, cohesive directory may remain flat when its files are true peers and further grouping would add navigation without clarifying ownership. A reviewer should block new placement that enlarges a heterogeneous flat root when a stable domain or test-layer boundary is already evident.
