+++ package.json
@@ -45,12 +45,12 @@
   "scripts": {
     "dev": "bun run src/cli/main.ts",
     "doctor": "bun run src/cli/main.ts doctor",
-    "docs:check:staged": "bun run scripts/check-staged-markdown-wrap.ts",
-    "skills:docs:sync": "bun run scripts/sync-bundled-skill-references.ts",
-    "skills:docs:check": "bun run scripts/sync-bundled-skill-references.ts --check",
-    "smoke:malformed-tool-pty": "bun run scripts/smoke-malformed-tool-pty.ts",
-    "smoke:tmp-scratch-pty": "bun run scripts/smoke-tmp-scratch-pty.ts",
-    "smoke:skillify-pty": "bun run scripts/smoke-skillify-pty.ts",
+    "docs:check:staged": "bun run scripts/check/check-staged-markdown-wrap.ts",
+    "skills:docs:sync": "bun run scripts/check/sync-bundled-skill-references.ts",
+    "skills:docs:check": "bun run scripts/check/sync-bundled-skill-references.ts --check",
+    "smoke:malformed-tool-pty": "bun run scripts/smoke/smoke-malformed-tool-pty.ts",
+    "smoke:tmp-scratch-pty": "bun run scripts/smoke/smoke-tmp-scratch-pty.ts",
+    "smoke:skillify-pty": "bun run scripts/smoke/smoke-skillify-pty.ts",
     "hooks:install": "git config core.hooksPath .githooks",
     "lint": "biome lint .",
     "typecheck": "tsc --noEmit",
@@ -67,14 +67,14 @@
     "test:acceptance:responses:deepseek": "bun test ./tests/acceptance/provider/deepseek-responses.acceptance.ts",
     "test:acceptance:mcp": "bun test ./tests/acceptance/mcp/real-servers.test.ts",
     "prepack": "bun run build:npm",
-    "build:npm": "bun run scripts/build-npm-package.ts",
-    "pack:check": "bun run scripts/check-npm-pack.ts",
-    "pack:smoke": "bun run scripts/smoke-npm-package.ts",
-    "smoke:binary": "bun run scripts/smoke-binary.ts",
-    "build:exe": "bun run scripts/build-exe.ts",
-    "build:assets": "bun run scripts/build-assets.ts",
-    "build:installer:stage": "bun run scripts/stage-windows-installer.ts",
-    "build:installer": "bun run scripts/build-installer.ts"
+    "build:npm": "bun run scripts/build/build-npm-package.ts",
+    "pack:check": "bun run scripts/check/check-npm-pack.ts",
+    "pack:smoke": "bun run scripts/smoke/smoke-npm-package.ts",
+    "smoke:binary": "bun run scripts/smoke/smoke-binary.ts",
+    "build:exe": "bun run scripts/build/build-exe.ts",
+    "build:assets": "bun run scripts/build/build-assets.ts",
+    "build:installer:stage": "bun run scripts/build/stage-windows-installer.ts",
+    "build:installer": "bun run scripts/build/build-installer.ts"
   },
   "dependencies": {
     "@modelcontextprotocol/client": "2.0.0",
