+++ scripts/check/check-npm-pack.ts
