+++ scripts/smoke/smoke-npm-package.ts
@@ -1,7 +1,7 @@
 import { mkdir, mkdtemp, readFile, readdir, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import packageJson from "../package.json";
+import packageJson from "../../package.json";
 
 export {};
