+++ scripts/smoke/smoke-windows-installer.ps1
