+++ scripts/smoke/support/pty-smoke.ts
